requirements.txt
numpy==1.21.4
pandas==1.3.5
matplotlib==3.5.1
sklearn
seaborn
progressbar2
jupyter
psutil
